fix: address second round of CodeRabbit review
Collab authorization (critical):
- JoinProject now requires the socket to already be a participant of the
  project's room; otherwise it returns PermissionDenied rather than
  allocating a replica for a guessed project id.
- Project-scoped broadcasts (UpdateBuffer / UpdateWorktree /
  UpdateRepository / UpdateDiagnosticSummary / UpdateLanguageServer)
  and every host-bound forward (SaveBuffer, OpenBuffer*, LspRequest,
  Git*) now drop/refuse messages from non-collaborators via
  CollabHub::is_project_collaborator.
- UpdateParticipantLocation is gated on room membership.

Collab hygiene:
- deregister() now tears down every project hosted by the departing
  connection and returns a DisconnectEffects summary; ws.rs broadcasts
  UnshareProject to the remaining guests so they learn the host is gone
  instead of racing against stale state.
- Outbound channel is bounded (capacity 128). Slow clients are
  disconnected on TrySendError::Full instead of allowing unbounded
  memory growth.
- ShareProject rejects with NotFound when the caller isn't already in
  the target room, and now persists sp.worktrees / sp.is_ssh_project
  in ProjectState so JoinProjectResponse returns them to guests.
- forward_to_host checks connection_alive for the host and surfaces a
  Disconnected error to the sender instead of silently dropping.
- Hello now rejects invalid tokens (PermissionDenied) instead of
  silently downgrading the session to anonymous. No token present is
  still an explicit anonymous session.

CLI / core:
- Global `--server` / `YUXU_SERVER` flag on yuxu-cli bootstraps a
  first connection to a remote backend without editing config.toml.
- Claims::exp / ::iat switched from usize to u64 to avoid 2038-era
  truncation on 32-bit targets.

+++ crates/yuxu-core/src/auth/jwt.rs
@@ -7,8 +7,8 @@ pub struct Claims {
     pub sub: String,
     pub username: String,
     pub is_admin: bool,
-    pub exp: usize,
-    pub iat: usize,
+    pub exp: u64,
+    pub iat: u64,
 }
 
 #[derive(Clone)]
@@ -34,8 +34,8 @@ impl JwtService {
             sub: user_id.to_string(),
             username: username.to_string(),
             is_admin,
-            iat: now.timestamp() as usize,
-            exp: exp.timestamp() as usize,
+            iat: now.timestamp() as u64,
+            exp: exp.timestamp() as u64,
         };
         Ok(encode(&Header::default(), &claims, &self.encode)?)
     }
